plugin: move get_queue_info () to accounting.cpp
Problem: get_queue_info () is a helper function that is specific to
flux-accounting in the fact that it is validating a queue that a user
may or may not have submitted a job under and potentially receiving a
priority bump because of it. It would make more sense to move this out
of the plugin and into accounting.cpp.

Move get_queue_info (), struct queue_info, and the special queue
definitions to accounting.*. Add the "queues" map as a parameter to the
get_queue_info () definition.
